+++ src/main/java/de/skuzzle/ghpromexporter/appmetrics/AppMetrics.java
@@ -1,5 +1,6 @@
 package de.skuzzle.ghpromexporter.appmetrics;
 
+import io.micrometer.core.instrument.DistributionSummary;
 import io.micrometer.core.instrument.Metrics;
 import io.micrometer.core.instrument.Timer;
 
@@ -12,8 +13,13 @@ public final class AppMetrics {
     private static final String NAMESPACE = "ghp_";
 
     private static final Timer scrapeDuration = Metrics.timer(NAMESPACE + "repository_scrape_duration");
+    private static final DistributionSummary registeredScrapers = Metrics.summary(NAMESPACE + "registered_scrapers");
 
     public static Timer scrapeDuration() {
         return scrapeDuration;
+    }
+
+    public static DistributionSummary registeredScrapers() {
+        return registeredScrapers;
     }
 }