+++ src/main/java/de/skuzzle/ghpromexporter/scrape/MemoryRegistrationRepository.java
@@ -2,12 +2,13 @@ package de.skuzzle.ghpromexporter.scrape;
 
 import java.util.Set;
 import java.util.concurrent.ExecutionException;
+import java.util.function.Function;
 
-import com.google.common.base.Supplier;
 import com.google.common.cache.Cache;
 
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
+import reactor.core.scheduler.Schedulers;
 
 class MemoryRegistrationRepository implements RegistrationRepository {
 
@@ -18,14 +19,20 @@ class MemoryRegistrationRepository implements RegistrationRepository {
     }
 
     @Override
+    public int estimatedCount() {
+        return (int) registeredScrapers.size();
+    }
+
+    @Override
     public boolean isEmpty() {
-        return registeredScrapers.size() == 0;
+        return registeredScrapers.size() == 0L;
     }
 
     @Override
     public Flux<RegisteredScraper> registeredScrapers() {
         final Set<RegisteredScraper> scrapers = Set.copyOf(registeredScrapers.asMap().keySet());
-        return Flux.fromIterable(scrapers);
+        return Flux.fromIterable(scrapers)
+                .subscribeOn(Schedulers.boundedElastic());
     }
 
     @Override
@@ -44,10 +51,11 @@ class MemoryRegistrationRepository implements RegistrationRepository {
     }
 
     @Override
-    public Mono<RepositoryMetrics> getExistingOrLoad(RegisteredScraper scraper, Supplier<RepositoryMetrics> loader) {
+    public Mono<RepositoryMetrics> getExistingOrLoad(RegisteredScraper scraper,
+            Function<RegisteredScraper, RepositoryMetrics> loader) {
         return Mono.fromSupplier(() -> {
             try {
-                return registeredScrapers.get(scraper, () -> loader.get());
+                return registeredScrapers.get(scraper, () -> loader.apply(scraper));
             } catch (final ExecutionException e) {
                 throw new RuntimeException(
                         "Error while loading fresh metrics after cache miss for " + scraper, e);