+++ src/main/java/de/skuzzle/ghpromexporter/scrape/RepositoryMetrics.java
@@ -1,5 +1,10 @@
 package de.skuzzle.ghpromexporter.scrape;
 
+/**
+ * Raw results of a single repository scrape.
+ *
+ * @author Simon Taddiken
+ */
 public record RepositoryMetrics(
         long totalAdditions,
         long totalDeletions,
@@ -9,4 +14,6 @@ public record RepositoryMetrics(
         int subscriberCount,
         int watchersCount,
         int sizeInKb,
+
+        /* Note: this field must always occur last */
         long scrapeDuration) {}